@philnik777 2024-08-17
Couldn't we simply forward to `std::sort` for these cases here?