@izvolov 2024-08-17
Radix sort is much faster than comparison sorts (including `std::sort`), especially in case of random input data. So forwarding to `std::sort` will not provide desired speed up.
[Here are some old benchmarks of raw radix sort in my library](https://github.com/izvolov/burst?tab=readme-ov-file#%D1%86%D0%B5%D0%BB%D0%BE%D1%87%D0%B8%D1%81%D0%BB%D0%B5%D0%BD%D0%BD%D1%8B%D0%B5-%D1%81%D0%BE%D1%80%D1%82%D0%B8%D1%80%D0%BE%D0%B2%D0%BA%D0%B8).