@izvolov 2024-08-17
Oh, I didn't understand the question first time.

The key moment is that radix sort needs a temporary buffer, and `std::stable_sort` has one, but `std::sort`, unfortunately, not.