@izvolov 2024-11-10
> The latest CI failures were unrelated to this PR.
> 
> The one of (generic-hardening-mode-fast-with-abi-breaks, libcxx-runners-set) is being fixed by #115271.

Thanks for making it clear. Should I wait for that PR merged?