@izvolov 2024-12-11
> Should code in `stable_sort.pass.cpp` be reformatted , or maybe ignored using `// clang-format off`?

Nearby tests are _almost_ formatted correctly, so I reformatted the test.